🚚 Footer directoryを作成してファイルをまとめました
